+++ lib/speed.js
@@ -1,33 +1,34 @@
 'use strict';
 
 const fs = require('fs');
-const { fork } = require('child_process');
+const { fork, execSync } = require('child_process');
 const makeResults = require('./speed/make-results.js');
 const PATH_TO_GET_TIME = __dirname + '/speed/get-time.js';
-const GET_NODE_PATH = __dirname + '/speed/get-node-path.sh';
+const PATH_TO_GET_NODE_PATH = __dirname + '/speed/get-node-path.sh';
+
+const DEFAULT_COUNT = 10000;
 
 const getNodePath = (
-  // Function which get path to node by version
+  // Function which gets path to node by version
   version // String, versions of node, example: 9.4.0, 8.0, 7
   // Returns: String, path to node
 ) => {
-  const execSync = require('child_process').execSync;
   let pathToNode = null;
   if (process.env.NVM_DIR) {
-    pathToNode = execSync(['sh', GET_NODE_PATH, version, 'nvm'].join(' '))
-      .toString().trim();
+    pathToNode = execSync(['sh', PATH_TO_GET_NODE_PATH, version, 'nvm']
+      .join(' ')).toString().trim();
     if (fs.existsSync(pathToNode)) return pathToNode;
   }
   if (process.env.NVS_HOME) {
-    pathToNode = execSync(['sh', GET_NODE_PATH, version, 'nvs'].join(' '))
-      .toString().trim();
+    pathToNode = execSync(['sh', PATH_TO_GET_NODE_PATH, version, 'nvs']
+      .join(' ')).toString().trim();
     if (fs.existsSync(pathToNode)) return pathToNode;
   }
   throw new Error('Node: ' + version + ' not found');
 };
 
 const getFullVersion = (
-  // Function which get full node version from node path
+  // Function which gets full node version from node path
   pathToNode // String, path to node
   // Returns: String, full node version
 ) => {
@@ -36,8 +37,8 @@ const getFullVersion = (
   return version[0] === 'v' ? version.substring(1, version.length) : version;
 };
 
-const makeRequests = (
-  // Function which create requests to child processes
+const prepareRequests = (
+  // Function which prepares requests to child processes
   versions, // Array of string, array of node versions
   testsSync, // Array of synchronous functions
   testsAsync // Array of asynchronous functions, callback-last
@@ -69,10 +70,10 @@ const getPathFromStack = (
 };
 
 const speed = (
-  // Function which manage processes of testing functions
+  // Function which manages the process of testing functions
   caption, // String, caption of test
   testFunctions, // Array, sync functions and array of async functions
-  params = {} // Object, parametrs (count, versions, MAX_ANOMALY_PERCENT)
+  parameters = {} // Object, parametrs (count, versions, MAX_ANOMALY_PERCENT)
 ) => {
   const syncFunctions = []; // Array of synchronous functions
   const asyncFunctions = []; // Array of asynchronous functions, callback-last
@@ -89,16 +90,14 @@ const speed = (
     return;
   }
   const path = getPathFromStack();
-  const DEFAULT_COUNT = 10000;
-  const count = params.count || DEFAULT_COUNT;
-  const MAX_ANOMALY_PERCENT = params.MAX_ANOMALY_PERCENT || 5;
-  const versions = params.versions || [process.versions.node];
+  const count = parameters.count || DEFAULT_COUNT;
+  const MAX_ANOMALY_PERCENT = parameters.MAX_ANOMALY_PERCENT || 5;
+  const versions = parameters.versions || [process.versions.node];
 
-  const results = new Map(); // results of all functions of all node versions
-  results.caption = caption;
-  results.count = count;
+  const results = new Map();
+  // results of all functions on every requested node version
 
-  const requests = makeRequests(versions, syncFunctions, asyncFunctions);
+  const requests = prepareRequests(versions, syncFunctions, asyncFunctions);
   const numberOfFuncs = syncFunctions.length + asyncFunctions.length;
 
   const sendRequest = (
@@ -116,7 +115,9 @@ const speed = (
       versionResults.push(result);
       if (versionResults.length === numberOfFuncs) {
         results.set(version, versionResults);
-        if (results.size === versions.length) makeResults(results);
+        if (results.size === versions.length) {
+          makeResults(results, caption, count);
+        }
       }
       if (requests.length) sendRequest(...requests.pop());
     });
