+++ lib/speed/make-results.js
@@ -3,13 +3,15 @@ const fs = require('fs');
 
 const outputToConsole = (
   // Function which log results to console
-  result // Map, sorted results of all functions of all node versions
+  result, // Map, results of all functions of all node versions
+  caption, // String
+  count // Number
 ) => {
   const rpad = (s, count) => (s + '.'.repeat(count - s.length));
   const lpad = (s, count) => ('.'.repeat(count - s.length) + s);
 
-  console.log('\n' + rpad(result.caption, 35) +
-    lpad('Count: ' + result.count, 15));
+  console.log('\n' + rpad(caption, 35) +
+    lpad('Count: ' + count, 15));
 
   result.forEach((times, version) => {
     console.log('\n' + lpad('Node version: ' + version, '.', 70));
@@ -28,12 +30,12 @@ const outputToConsole = (
 
 const writeToCsv = (
   // Function which write results to csv file
-  result // Map, sorted results of all functions of all node versions
+  result // Map, results of all functions of all node versions
 ) => {
   const filename = __dirname + '/results/data.csv';
 
-  const caption = 'funcName,time,percent,version\n';
-  fs.writeFileSync(filename, caption);
+  const captionOfFile = 'funcName,time,percent,version\n';
+  fs.writeFileSync(filename, captionOfFile);
 
   result.forEach((times, version) => {
     times.forEach(func => {
@@ -46,7 +48,9 @@ const writeToCsv = (
 
 const makeResult = (
   // Function which generate result of test
-  results // Map, results of all functions of all node versions
+  results, // Map, results of all functions of all node versions
+  caption, // String
+  count // Number
   // Returns: Map, sorted results of all functions of all node versions
 ) => {
   results.forEach((result, version) => {
@@ -58,7 +62,7 @@ const makeResult = (
     });
     results.set(version, sortedResults);
   });
-  outputToConsole(results);
+  outputToConsole(results, caption, count);
   writeToCsv(results);
   return results;
 };
